refactor(vite-plugin-angular): extract tailwind plugin into separate file
Move the tailwind-reference inline plugin, buildStylePreprocessor, and
validateTailwindConfig into tailwind-plugin.ts. The duplicate analog()
registration guard is kept in angular-vite-plugin.ts as it is not
tailwind-specific.
